discovery/requirements.txt
requests>=2.31.0
python-jobspy>=1.1.80
